prjfs: do not send duplicated directory entries on enumeration
Summary:
There was a subtle bug in D31782916 (https://github.com/facebookexperimental/eden/commit/87ef50135ec805ac027bc040032a254db91b32c1) where a wildecarded enumeration may send
the same directory entry multiple times to ProjectedFS. This can happen if the
first directory entry returned to ProjectedFS is not where iter_ initially
points to, which is almost always the case when a wildecard is being used. Thus
the first call to advanceEnumeration will simply find the first element sent to
ProjectedFS and send it again on the next readdir call.

Reviewed By: fanzeyi

Differential Revision: D32742433

fbshipit-source-id: 12904a5d2df4407072b8d775950f92ea93344a73