mononoke: add admin command to rederive commits
Summary:
We have the same command in backfill_derived_data, but backfill_derived_data is
not installed on devservers and few people know about this binary. Let's add
an option in mononoke_admin to derive or rederive commits. I want to have this
option because I'm going to enable new parallel derivation soon, and I want to
have a quick and easy way to rederive commits in case they get corrupted.

Reviewed By: markbt

Differential Revision: D32696853

fbshipit-source-id: 782e721115edf38a7a7b81673a8a52a160c7225f