Add --max-untracked-size option to create
Summary:
## What
This diff adds a `--max-untracked-size` option to `snapshot create` that allows skipping large untracked files.

## Why
It might happen that a large untracked file is created under the repository, and the use case might not really need to record those in the snapshot. This has been raised by some CiP use cases, like ephemeral fbpkg auditing.

## How
This just skips the files when uploading. It **does not** skip tracked files (modified files and files which were `hg add`ed).

In the future we might be more elaborate about this, and record in the snapshot that the file was skipped, but let's start with this easier take.

Reviewed By: markbt

Differential Revision: D32724049

fbshipit-source-id: f7af1a5c973efc5b1543c6598c8f21dcd777311a