manifest-tree: always create diff progress bar
Summary:
Previously the progress bar was only rendered via nativecheckout, but the tree-manifest is used in non-nativecheckout as well.

Also, set the total progress value to 18 to give a real progress bar instead of a spinner. This isn't the actual theoretical max, but everything after 15 goes very quickly and this is the highest depth I've seen displayed.

Reviewed By: DurhamG

Differential Revision: D32512924

fbshipit-source-id: d3bfe1e119ca7a416db11dc5cc0dedd101e7289d