progress: display total bytes in time series
Summary: Now the HTTP time series progress will also display the total downloaded/uploaded bytes. This is useful to gauge if we are transferring a reasonable amount of data. A further improvement might be to otuput the total when the command finishes so it is available even if the user wasn't watching the progress.

Reviewed By: quark-zju

Differential Revision: D32736856

fbshipit-source-id: d6a9391b66bf5fe686770c5058c813ed20050fc7