Teach the rest of segmented changelog to use the `dag` crate
Summary:
We've been running this successfully in the tailer; now switch over the rest of the code base.

Note that for on demand segmented changelog builds (in-memory only), we can end up desynced - we load an IdMap that knows about some commits, but the IdDag does not. To resolve this, the update code checks to see if the IdDag knows about the head commit we've just asked for (which implies that all ancestor commits are in the IdDag, too). If it does not, we use the slow path to rebuild part of the IdDag from the IdMap.

Reviewed By: quark-zju

Differential Revision: D32641483

fbshipit-source-id: c554b6e6f38bb9853a24bc29e8a60e403beb4cd2