integration: fix test_file_locked_removal on Windows
Summary:
The test started failing on CI due to the file not being read when being
locked. ProjectedFS behavior when the file is a placeholder vs a full file
differs when that file is locked and invalidated. In the case of a placeholder,
the file will be removed, while the full file won't. Let's force a full file by
simply reading it before locking it.

Reviewed By: kmancini

Differential Revision: D32776736

fbshipit-source-id: faf0bca1becc4bfed8cd6bc804e33a0540f3b8cb