Change derived data config to sets of configs
Summary:
To be able to handle requsets from clients with different configs on the derived data service let's make a map Config_name -> Config and send the config name for the config we want service to use within the request.
I've added new fields in D32460282. This diff replace consumption of the new config.

Reviewed By: mitrandir77

Differential Revision: D32463274

fbshipit-source-id: ca125c231341a7f98391800ca19cab5ea8920dd6