tidy up commit messages
Summary:
We had TODOs to figure out the right commit metadata here. I think what we
have right now is mostly fine as long we keep it consistent between commits.

Centralizing will also make it easier to change it in the future if we need it.

I've also unified capitalization for those.

Reviewed By: StanislavGlebik

Differential Revision: D32468277

fbshipit-source-id: 41d3286d9e33e39741c1f8196316f8dfcfc394db