resolve: record dirstate updates after mergedriver runs
Summary:
mergedriver may queue adds/removes in the mergestate. These are
processed by ms.recordactions() to update the state of the files. Previously
recordactions was happening before the mergedriver ran, so any adds/removes it
added wouldn't actually get recorded. This doesn't happen in the normal 'hg
rebase' case because it has it's own resolution path that is different from 'hg
resolve'.

This is most notable in VS Code, which runs the initial rebase as
--noninteractive, then handles all resolves via hg resolve. This bug has been
here since 2015, so I have no idea how this hasn't been noticed as a major
breakage before.

Reviewed By: quark-zju

Differential Revision: D32628482

fbshipit-source-id: c60745d608b39c827a520712aa20756be6d61071