cli_rs: propagate edenfsctl.real.exe error code to edenfsctl
Summary:
Some scripts are using the return status of edenfsctl to decide whether to
continue or abort their operation. We need to make sure that when an error is
thrown by edenfsctl.real.exe, the Rust CLI exit with the same error code to
make sure that these scripts continue working as expected.

Reviewed By: skotchvail

Differential Revision: D32818993

fbshipit-source-id: 21df2153d469d3a48bfba046a07360726498806c