scsc: add blame --parent-index option
Summary:
The `blame` subcommand has a `--parent` option for blaming the parent of an
identified commit, rather than the target commit.  This can be useful when
skipping over a commit, as it avoids the need to look-up the parent.

It currently only works for the first parent, which most of the time is
sufficient.  However, in the case of a merge where the blamed line came from
another parent, we need to be able to specify a different parent.  This is a
corner case, but it will be more visible with blame parents now being visible
(the parent index is provided in the response from the server in the
`parent_index` field).

With this change, it's possible to get a high-quality skip-past-this-commit for any line in a single step by using:
* `--commit-id` with the commit identity of the blamed line
* `--parent-index` with the parent index of the blamed line
* seeking to the range described by `parent_start_line` and `parent_range_length` of the blamed line.

Reviewed By: yancouto

Differential Revision: D32617024

fbshipit-source-id: 0778b187a8610ac3f5f53666a0cba023b32455ac