blame_v2: provide blame parent information to clients
Summary:
Blame V2 allows us to provide information about the range of lines in the
parent commit that were replaced by a line.  Include this information in the
response to clients (if requested).

Extend `scsc` with `--parent-line-range` to allow display of this value.  We
also add `--origin-path` so that we can see the original path, too.

Reviewed By: yancouto

Differential Revision: D32616108

fbshipit-source-id: 5050e7e09d255ca2287859ad270f67f87c954cf4