Add --max-untracked-size option to latest
Summary:
See D32724049 (https://github.com/facebookexperimental/eden/commit/ca7e6208f07abb62ab4ec7ee17dd99300d739011) for summary of how this works when creating commits.

This diff implements the "latest" counterpart, which aims to get rid of this possible inconsistency:
```
> create some large file
> hg snapshot create
> hg snapshot latest --is-working-copy
FAIL
```

where `latest --is-working-copy` will say the snapshot is not the working copy when in fact it was created from it. With this diff as long as you specify the same maximum untracked size, the results should be consistent.

Reviewed By: StanislavGlebik

Differential Revision: D32726009

fbshipit-source-id: c438155556d578e08b1084ad3065b1b4700503fb