mononoke: support separate bookmarks for source repos and hyper repo
Summary:
Let's make it possible to support different bookmarks for source repos and
hyper repo. The main reason for that is to make it harder to mistakenly sync
from a test repo into a prod repo. E.g. if bookmarks are the same then
reversing the direction of the sync by mistake is possible and that could lead
to prod repo corruption.

If source and hyper repo bookmarks are different then even if we mistake the
direction of the sync the hyper_repo_builder will immediately fail without
creating any repo corruptions.

Reviewed By: yancouto

Differential Revision: D32790791

fbshipit-source-id: 9d482b0ab3cdbd3c925a0afdd573af7f2382f5ae