mononoke: add tail command to hyper repo builder
Summary:
This diff adds tailing to hyper-repo i.e. the ability to mirror commits from
source repos into hyper repo.
Two biggest limitations so far is that non-forward bookmark moves are not
supported and that syncing merge commits is not supported. The former
limitation probably won't be lifted, while the latter will be.

Reviewed By: yancouto

Differential Revision: D32755551

fbshipit-source-id: 4d813024a2330e7843d898cd1d5db8cbef3eace9