mononoke: introduce hyper repo builder
Summary:
This binary is a tool that can be used to create a test repo from a collection
of real prod repos.
The main goal is to stress-testing derive data derivation - by combining large
repos together we get a repo with a very high commit rate, and we can check if
derived data derivation can keep up with that rate.

This tool allows us to import a working copy of a given source repo into a
"hyper-repo", and then tail new commits that land in a source repo. Note that
it doesn't import any of the history, just the working copy itself.

This diff adds a command that let's us add a new source repo into existing
hyper-repo. All files from the source repo go to the directory of the same in
the target repo e.g. files for source repo "source" will be in directory
"source/..." in target repo.

What might be the most controversial decision is how to track which were the
latest commits that were replayed to the hyper-repo from the source repo. I
decided to track it in commit extras. Initially I wanted to have a separate
folder in the repo that store latest synced commit for every repo, however that
would mean that in order to find the latest replayed commit we'd need to derive
a manifest. I wanted to avoid doing that since I want derivation to be done by
a completely different process (derived_data_tailer) to make measurements
simpler, hence opted for commit extras.

Reviewed By: yancouto

Differential Revision: D32755550

fbshipit-source-id: d42acabc8acb561c4ca243a91d9239c04e1a0fb8