mononoke: fail hyper_repo_builder if a repo has a backup repo
Summary:
I want to be extra cautious, so I suggest to fail if a given hyper repo has a
backup config. Since hyper repo is expected to be a test repo it would be
unusual for it to have a backup config.

We might later decide that this is a weird restriction and lift it, but for now
I'd like to keep it.

Reviewed By: yancouto

Differential Revision: D32794923

fbshipit-source-id: 3d21d16ccc79a326ed0b6b7901c91dc601ce7312