Do bulk fetches in the segmented changelog tailer
Summary:
When we're a long way behind, a bulk fetch is faster. When we're not far behind, it should cost similar I/O to fetching a single commit directly.

Use bulk fetches always, so that whenever we get a long way behind, we tail nicely, without penalty for when we're near the tip

Reviewed By: StanislavGlebik

Differential Revision: D32641784

fbshipit-source-id: 0f68d284b1b579ce78e8946d9baaafac0e965071