edenapi: don't retry if we hit tls errors
Summary:
Eden is seeing issues where checkouts take forever because every
request to edenapi hangs for multiple seconds while we hit tls errors, then
sleep 1s, then retry. Tls errors are not going to go away after a single retry,
so let's not mark them retryable.

In theory eden should fail in that case, and raise the cert error to the user
attention, but for some reason falling back to the importhelper succeeds, which
implies the cert is not actually bad. So there's more to debug here, but this
should stop the bleeding.

Reviewed By: xavierd

Differential Revision: D32934250

fbshipit-source-id: f2ff2fdeffb55f2f582560d09cfefe9b925dd91b