mononoke: make DerivationContext::flush() call consume itself
Summary:
### Description of the problem

MemWritesBlobstore::persist() has a footgun. When persist() call is made it
first cleans the cache and only then it persists the data to underlying
blobstore. This can lead to the race condition if MemWritesBlobstore
used concurrently by a few tasks. For example

1) Two tasks read and write data to MemWritesBlobstore concurrently. So far
everything is fine
2) Then one of the the tasks call `persists()` and cleans the cache, but didn't
manage to save the data to blobstore yet
3) When the cache is already clean but then data is not persisted yet another
task tries to read the data that has just been in the cache. Because cache is
clean it fails to find the blob.

I believe we didn't really have this problem in production, but it was the case only
because we never used `enabled_write_batching()` outside of batch derivation.
But if we used `enabled_write_batching()` in non-batch derivation, then because
of the `flush()` calls we had in `perform_single_derivation_locally()` function
it would have ran into race condidtion described above.

### Attempts to remove the footgun

I've made a few attempts to make it harder to shoot yourself in the foot i.e.
make it harder or impossible to have owners of the same MemWritesBlobstore that
can call persist at any time.

Below I document my attempts to fix it

1) In D32595096 I made an attempt to fix this issue by making `persist()` consume
MemWritesBlobstore object, but this was a bad idea. We do a lot of copies of
MemWritesBlobstores, and we also put these copies into `Arc`. Because of that
when we call `persist()` we consume one of the MemWritesBlobstore object,
but there might be other clones of the same objects that are still alive and
reference the same cache.
This solution could have worked if we made MemWritesBlobstore object
non-cloneable, and then only passed `&dyn Blobstore` to derived data. However
this is not really feasiable - derived data objects do `spawn()` which require
blobstore to be `Arc`ed.

2) Another option was to make `DerivationContext::flush()` consume the object AND
make `DerivationContext` non-cloneable. If we had that then calling `flush()` requires you to
own the DerivationContext object, which hopefully would make it clearer that MemWritesBlobstore
requires special care.
Unfortunately making DerivationContext non-cloneable is hard. The main reason is because
DerivationContext is passed to a `tokio::spawn`ed tasks (see blame_v2), which requires it to
be cloneable.

### This diff

So this diff makes `DerivationContext::flush()` call consume
itself. It's still easy to bypass it by doing `derivation_ctx.clone().flush()`, but hopefully
it should be enough to bring the attention.

Reviewed By: farnz

Differential Revision: D32638849

fbshipit-source-id: 3bb61f0e29d0fccb818e0936c7a68eb5ddb2f921