phases: remove abstraction violation
Summary:
The `Phases` trait has a method that returns its one and only implementation:
`SqlPhases`.  This means we can't reimplement the trait, and we also can't use
it to break the dependency graph, as the trait itself has its implementation
as a dependency.

Remove this dependency by expanding the `Phases` trait to include
methods that do the functionality that client code was relying on when it
called `.get_store()`.  This is the de-facto interface for Phases anyway, so
we may as well make it official.

We will use this to split the interface from the implementation.

Reviewed By: yancouto

Differential Revision: D32678847

fbshipit-source-id: 3c27e2fb41ccfd69ee2af2416c732e77bf5a130a