cli: properly implement is_process_alive on Windows
Summary:
On Windows, even when a process is terminated, its handle can still be queried,
for instance to get its exit code. Thus merely checking if the handle can be
opened isn't a good way to test if a process is alive. Let's instead query the
exit code for the process.

As a bonus, also removed the bit about EdenFS dropping root privileges since
EdenFS no longer starts as root on macOS or Linux.

Reviewed By: chadaustin

Differential Revision: D32995121

fbshipit-source-id: 9ac7946dbb53b3fabb163c98e3f72f9132b13a21