mononoke_api: fix commit history parameter precedence
Summary:
If both `descendants_of` and `exclude_changeset_and_ancestors` are specified,
and both are set to the same commit, then that commit should be excluded (i.e.
the `exclude_changeset_and_ancestors` parameter should take precendence.

While we are here, improve the documentation so the purpose of these parameters
is clearer.

Reviewed By: bolinfest

Differential Revision: D32991576

fbshipit-source-id: aafb65b58ac2cea1dce403fa61ddcfdd6e6becfd