add nested checkout restrictions to eden clone
Summary:
By default, we do not want to support nested checkouts. Their behavior is not well defined and can cause issues.

This diff disallows nested checkouts unless specifically requested by users. They can use the --allow-nested-checkout or -n flags to force nested checkout creation. If these flags are not used, the clone is aborted and an error message is printed.

Reviewed By: chadaustin

Differential Revision: D32514269

fbshipit-source-id: d04a7ef82c6d31993801648da42e4182167f37e6