manifest: add `after` parameter to ordered traversal
Summary: When specified, the `after` parameter will cause traversal to include only those paths which are after the given path in the ordering of the repository.  This will allow us to implement continuation of ordered traversal based on the value of the last path that was reached in a previous traversal.

Reviewed By: StanislavGlebik

Differential Revision: D32312594

fbshipit-source-id: 973a04350c8e0f91c18dda7630ae18c4eab3d74c