mononoke: log to ods if a new parallel derivation is used
Summary:
The main motivation is to just know for sure that a new derived data
derivation is used - otherwise it's hard to tell.

Initially I wanted to log to scuba but it turned out it to be trickier than I
thought. Since we need it only for the rollout phase (which hopefully should be
pretty short), let's just log to ods.

Reviewed By: yancouto

Differential Revision: D32956728

fbshipit-source-id: c71aa3dfef52a09314a7fb8d0e440706536b7731