don't validate the source bookmark
Summary:
Those bookmarks don't have to exist in the source repo. Let's skip this
validation and assume that our client knows what it's doing.

We already check that the commit we were requested to sync is the child
of the last commit we synced from that source which is a pretty strong
validation.

Context: we dropped the bookmarks and currently the sync is blocked by this.

Reviewed By: mojsarn

Differential Revision: D33070449

fbshipit-source-id: fa689f1b2d63ea14f3f83473836aac49c16cb154