Support symlinks on create
Summary:
This is the first step to support symlinks on snapshots. As with usual changesets, this will be done by storing the path of the symlink as the file content (as well as storing that the file type is symlink).

This diff adds support for symlinks on "snapshot create":
- When creating a snapshot, if the file is a symlink, instead of reading its content, let's read what it points to and store that as file content.

Support on "snapshot update" is added on D32735123.

Reviewed By: ahornby

Differential Revision: D32735129

fbshipit-source-id: 1b462696e7ba6be9d8fa683ea5baf2af76687a5c