Fix missing file bug
Summary:
This was the error found in D32761279.

The fixed error is: When a file is marked as deleted but it is actually missing, the snapshot restore logic does not correctly mark it as missing again.

Unfortunately I didn't find a way to fix it without needing to write a dummy empty file.

Reviewed By: ahornby

Differential Revision: D32768443

fbshipit-source-id: 287bc0c33030dc8988f4cec8ea4b7b3c3342d417