Remove intermediate function for uploadsnapshot
Summary:
Right now, the flow is `python -> uploadsnapshot -*> uploadsnapshot_py -> upload_snapshot`.

The arrow marked with a `*` doesn't really do much. It is legacy from when `uploadsnapshot_py` did the bulk of the work, but now that behaviour has been moved to `upload_snapshot`, which knows nothing about python and thus is clearer. So this diff simply removes this intermediate step to simplify code a little.

Reviewed By: ahornby

Differential Revision: D32726077

fbshipit-source-id: 4b9b4700974ae38c9a7b443db12c8b3aa6265ee9