cli: remove unecessary `hg cat` when cloning
Summary:
On clone, the CLI tries to open an .arcconfig file at the given revision for
the solve purpose of printing a slightly different log message. Since this log
isn't necessary to be present, let's avoid the `hg cat` and simplify the code a
bit on top of making it faster.

This should make cloning a couple of seconds faster on Windows due to Mercurial
taking roughly a second to start.

Reviewed By: chadaustin

Differential Revision: D33070099

fbshipit-source-id: 81c5cba8348d0c8d82c57a6656e6eb79c2b1f878