win: test if process is alive in is_edenfs_process
Summary:
On Windows, even though a process died, its pid can still be queried, thus
issuing a `eden stop` followed by an `eden restart` would try to kill an
already dead process and fail with a permission denied that would be displayed
to the user.

Let's test if the process is still alive before claiming that a given PID is an
EdenFS process to solve this.

Reviewed By: genevievehelsel

Differential Revision: D33076917

fbshipit-source-id: 062ec9e99ba2014720e3aa8d873309f5e581185d