Extract common behaviour in create_changeset
Summary:
This simply extracts common behaviour to a method. The reason they can't be merged in the same function is that we have objects of two different types that implement the same traits, so we need to use some generics to be able to share code.

I'm extracting this because I plan to add more behaviour here and I don't want to duplicate it as well.

Reviewed By: mitrandir77

Differential Revision: D32994671

fbshipit-source-id: 74d8cf7edcc3534aabbd7e2d6c5fd6144493869e