Use reference on save_bonsai_changeset
Summary:
This simply changes `container: impl ChangesetsRef + RepoBlobstoreRef` to `container: &(impl ChangesetsRef + RepoBlobstoreRef)`, meaning we now need to pass a reference.

This is slightly more efficient, as we don't need to clone the arc, and will make it easier to make changes in next diffs, as we don't need to depend on the container being cloneable.

Reviewed By: mitrandir77

Differential Revision: D33068835

fbshipit-source-id: 49b2ef2986afc520f6c98eb5a777d1a13e9e0fba