Fail on ambiguous prev bottom
Summary:
# Context

See D32917522

# On this diff

This actually does not add interactive mode support anywhere, but it's a pre-requisite for doing so to `prev --bottom`. It finds all "candidate bottoms" (draft ancestor roots) for the current commit, and fails if there is more than one.

This is **a change to current behaviour**, which is to just choose the first of those roots. That's why I split it to a different diff.

I believe it's a correct change, given how hg behaves in all other scenarios (failing on ambiguity), but am unsure about how fine is it to change a behaviour in hg like this.

## Alternative:
If there are multiple possibilities, still default to using the first, and only on `--interactive` ask for which to pick. This would keep the current behaviour, but is inconsistent with the rest of `hg prev/next`, so I decided not to do that. If it is preferred, I can modify the diff to do that.

## Newest:
One thing I noticed is that `--newest` option is ignored on `hg prev --bottom`. We should probably either fail if they're both present, or make it work, which would need some small changes to `_findprevtarget` to allow it to iterate until we're at the bottom draft commit.

Reviewed By: quark-zju

Differential Revision: D32921762

fbshipit-source-id: bc28c05e69f421b767709c9638a29d90b90a1111