interactive next
Summary:
# Context

This stack adds interactive mode to `hg next` (and `prev`). It provides the user with the option of what to do when there is more than one possible thing that could happen (which currently ends up in failure because of ambiguity).

I think this makes the command easier to use and more powerful (surely to me, plus this was initially suggested to me by victorsenam and other teammates thought it was a good idea).

The mode is default when the shell is interactive.

# On this diff

This diff implements interactive mode for the `hg next` command only. It sets up the groundwork for the other commands.

Reviewed By: quark-zju

Differential Revision: D32917522

fbshipit-source-id: 03328ca7b601804ea6f992820a4590a8d698dc58