configparser: use paths.default for fallback reponame
Summary:
The `reponame` is only used for various remote fetches. Requiring it makes
clone not "just work" - one has to pass something like
`--configfile /etc/mercurial/repo-specific/<repo>.hgrc` instead of just
`hg clone fb://fbsource`. It also makes the config system feels messy -
we have to ship the repo-specific config files and cannot just rely on
the dynamicconfig system to set everything up.

The ideal UX is that the user only need an URL like `fb://fbsource` and we
derive every other config from that single URL. This diff is a step torwards
that.

Note: there is a reason where `paths.default` fallback could be undesirable.
That is Sandcastle's uses of "master" local repos. There are local p2p clones
and the path might look like `/data/sandcastle/fbsource-with-random-suffix`.
However, that legacy setup is being removed, and we still have reponame set
correctly for them. So this change won't immediatly break anything (dropping
the `/etc/mercurial/repo-specific` config files might, but that's not this
diff).

Differential Revision: D32570352

fbshipit-source-id: d4b316e1a792aed1bd5b29233e9a8175fe798cdc