edenapi: use "reponame" associated by Client
Summary:
The EdenApi methods requiring `reponame`. It leaks the HTTP URL detail to many
more layers.

This diff makes the `Client` use the repo name provided by the config, or
explicitly via the `repo_name` API (as in tests). The HTTP client uses the repo
name in its config so callsites don't need to provide the name again.

Previously, if reponame is not provided, the HTTP client builder will use
"unknown" as fallback. That behavior is changed to an error instead.

The above change breaks some tests using production configs on scmstore
construction. scmstore is updated to properly avoid constructing EdenApi
if repo name or URL is not provided, instead of constructing an EdenApi
HTTP client using the wrong "unknown" URL silently.

This diff does not change the `EdenApi` trait yet, since that is a large change.

Reviewed By: yancouto

Differential Revision: D32800595

fbshipit-source-id: a156e18b2f6fc41804cc40e3c926086be63fd9b9