prevent hg doctor eden doctor cyclic loop
Summary:
I am about to teach eden doctor to run hg doctor in the backing repo for EdenFS
mounts to remediate some issues. hg doctor can call into eden doctor some times
and we do not want this to ever happen when eden doctor called hg doctor or we
will make a infinite loop of processes.

I decided to control this with an environment variable instead of an option on
hg doctor so that it won't show up in the help page for users.
credit: original idea to use env var from Genna

Reviewed By: xavierd

Differential Revision: D32971284

fbshipit-source-id: fc9925a132579ed01642eb35eac3eafc5f1e276d