modify NfsDirList to support entryplus3 entries
Summary:
The current readdir implementation uses NfsDirList to hold readdir results. Our new readdirplus implementation should use the same struct. To do this, we will add the ability to use an entryplus3 XdrList instead of an entry3 XdrList.

We also add a new constructor so we can specify which XdrList to initialize when we create an NfsDirList.

Reviewed By: xavierd

Differential Revision: D32708688

fbshipit-source-id: e2b324dd1dffc8620be89dca471043aba9b167a5