shelve: remove deprecated visibility code from test
Summary:
The shelve test used to have code to check visibility (with obsmarkers) worked
ok in a transaction.  We've removed obsmarkers, so this doesn't apply anymore.

Delete the test code - we've been running without it ok for a couple of years, and `shelve` is likely to get replaced by snapshots at some point anyway.

Reviewed By: DurhamG

Differential Revision: D33234288

fbshipit-source-id: 809e469016cd0c84927735f2f459e57743e90ec3