prune: make prune an alias for hide
Summary:
The `prune` command is left over from obsmarkers.  Previously, it was a way to manually construct obsmarkers.  Now it is functionally the same as `hg hide`.

There is a small difference: `hg prune` would move bookmarks to the parent,
rather than removing them when a bookmarked commit was hidden.  Prune is very
rarely used, so making this consistent seems ok.

Reviewed By: DurhamG

Differential Revision: D33235254

fbshipit-source-id: 7f688de2f60c13c172ce53f68a986a4cc94349fb