manifest-tree: make Entry::from_elements sort input
Summary:
Take some extra effort to force the right order when serializing entries to
bytes. Previously the order happens to match `BTreeMap<PathComponentBuf, _>`
order. But that is problematic for git.

This also makes `lookup_hg` able to use fast path to skip for tests that do
not have elements sorted.

Reviewed By: markbt

Differential Revision: D33369605

fbshipit-source-id: e939cc2089be63b69ccd590870b28f7468cb35ed