inodes: speedup overlay initialization on Windows by 15x
Summary:
In order for EdenFS to correctly invalidate all the on-disk placeholder and
full file that are present on disk, EdenFS relies on the loaded and unloaded
inodes. On startup, no inodes are loaded and thus EdenFS needs to fill up the
InodeMap with all the files present on disk. Unfortunately, doing so is very
expensive as checking if a file is present on disk is an IO operation, and thus
can take a really long time: we have seen users with several minutes InodeMap
initialization.

In order to reduce this time, EdenFS can try to be slightly less precise by
assuming that all files present in a on-disk directory have placeholders, and
thus not check them. This can cause more invalidation on update, but these are
much cheaper, and also concurrent, they are also not a bottleneck.

Since all directories on-disk must have an overlay entry, we also don't need to
check the working copy and just test the overlay.

Reviewed By: fanzeyi

Differential Revision: D33414475

fbshipit-source-id: efe0719bbac6cbba3aa4300fe4810dbee2e43841