run-tests: set FB_SCM_DIAGS_NO_SCUBA
Summary:
This affects the telemetry wrapper. Namely, the `git` binary will spawn a
`scm-telem-log` process which might have side effects writing `.scm.sqlite`
files when trying to log to scuba, which breaks the git tracking test in
`test-run-tests.t`.

A cleaner fix would be changing the git wrapper to stop spawning
`scm-telem-log` when `TESTTMP` is set. But that would take longer to roll out
and be effective.

Differential Revision: D33438494

fbshipit-source-id: d6af78be434cc0e966170dccaf4ff5c2b3868a52