cli: fix get_exit_code on Windows
Summary:
The _LPDWORD creates a nullptr, and thus GetExitProcess would fail to write to
it, raising an OSError which would pretend that the process is no longer alive.

This caused `eden stop` to not really wait for EdenFS to exit. Similarly,
`eden restart --force` would sometimes fail to start the new EdenFS due to the
old one not having released the lock file.

Reviewed By: genevievehelsel

Differential Revision: D33438953

fbshipit-source-id: 756ff4541ace5be0d081c8d4902838a50d2b134f