Add derivationType parameter to the request for derived data service.
Summary:
This and upcoming diffs are making possible to discover underived commits using derived data service. As a first step let's add the argument to derivation request, which denotes if it is a single commit derivation, or service should perform discovery of underived. Later we can expand that argument to cover batch/parallel derivation.
Clients should send requests derivation with discovery to the service, which in turns can distribute single derivation across different hosts using requests for single derivation.

Reviewed By: mitrandir77

Differential Revision: D32879222

fbshipit-source-id: 38fe38d92cb880d0cf1cfac230d39bb777c70f78