scs_server: provide last path for limited commit_compare
Summary:
When using paginated commit compare, it's not immediately obvious from the response whether the limit has been hit, and if so, which path should be used as the `after_path` parameter in the next request.

Make it easier by adding a new field in the response that is populated when the limit is hit, and can be taken as-is and passed into the next request.

Note that there is an edge case where the limit may be hit on the final path.  We will still return the `last_path` field in this case, and the next request will be empty.  This is unavoidable without scanning ahead, so clients will need to handle the possibility of an empty response.

Reviewed By: mitrandir77

Differential Revision: D33453953

fbshipit-source-id: 6ec95b51ddcdcb0fb5e75977deb944d0201fcb8f