Don't fail on parallel create
Summary:
This will fail a snapshot creation if two commands are run in parallel. We don't want this. Since writing latest to the metalog is of secondary importance, let's not fail the command if it fails.

In the future we want to always overwrite the metalog, but that doesn't seem so straightforward so I'll leave it to another diff.

We grab a lock when running create, but metalog is created when hg starts running, so it conflicts anyway. If we could reload after grabbing the lock, the issue would also be fixed.

Reviewed By: markbt

Differential Revision: D33168496

fbshipit-source-id: f4927c3bbca1a76669079b232d323636e56e8918