copytrace: treat missing copy sources as conflicts
Summary:
With fastcopytrace, attempting to rebase a directory rename over a file rename
(or vice versa) is not successful, as the copy source that fastcopytrace comes
up with doesn't exist in the rebase source commit.

Currently this crashes with an obscure `[copy source filename] not found in manifest`
error.  We can do better: if the copy source doesn't exist in the source
manifest at the point where we are merging, we can treat this as a conflict.
The user can either resolve this manually (by renaming the file to the new
destination), or they can try again with full copytrace, which should succeed.

While it's not strictly accurate, we treat this as a "change/delete" conflict,
as there is no "rename/delete" conflict type.

Reviewed By: DurhamG

Differential Revision: D33259386

fbshipit-source-id: 321f1942b0e31c3d97a4c4a32ee1eae6b6a740ce