tracing: add "sampling" tracing Layer
Summary: "sampling" refers to the python hg feature where certain ui.log keys can be marked for export to a specified file. The scm telemetry wrapper shuffles the file contents off to scuba. In rust, we now have a tracing Layer that implements the event export format. It matches tracing events using the "target" metadata attribute since that can be checked statically before the event is instantiated. Note that I am not currently taking advantage of that, but will in a following commit.

Reviewed By: quark-zju

Differential Revision: D32444379

fbshipit-source-id: c5d9fd5e28271656082d82f6584925b304ab02eb